Rename `config` module to `setup`. Split `ssp_setup.py` into separate modules.
`ssp_setup.py` split to:
- `logging.py`: logging setup
- `exit.py`: exit function
- `outdir.py`: functions to handle output directory
