+++ pyproject.toml
@@ -66,7 +66,7 @@ include = ["sourcespec2", "sourcespec2.*"]
 
 [tool.setuptools.package-data]
 "*" = ["LICENSE"]
-"sourcespec2.config" = ["*.yaml", "*.conf"]
+"sourcespec2.setup" = ["*.yaml", "*.conf"]
 "sourcespec2.html_report_template" = ["*.html", "*.css"]
 
 [tool.setuptools.dynamic]
