+++ sourcespec2/setup/mandatory_deprecated.py
