+++ sourcespec2/ssp_local_magnitude.py
@@ -23,7 +23,7 @@ from obspy.signal.filter import envelope
 from obspy.signal.invsim import WOODANDERSON
 from obspy.signal.util import smooth
 from obspy.signal.trigger import trigger_onset
-from .config import config
+from .setup import config
 from .ssp_data_types import SpectralParameter
 from .ssp_util import cosine_taper
 from .ssp_util import remove_instr_response
