+++ sourcespec2/ssp_plot_spectra.py
@@ -25,7 +25,7 @@ import matplotlib
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_pdf import PdfPages
 import matplotlib.patheffects as PathEffects
-from .config import config
+from .setup import config
 from .ssp_util import spec_minmax, moment_to_mag, mag_to_moment
 from .savefig import savefig
 from ._version import get_versions
