+++ sourcespec2/ssp_read_sac_header.py
@@ -13,8 +13,7 @@ import re
 import logging
 import contextlib
 from obspy.core.util import AttribDict
-from .config import config
-from .ssp_setup import ssp_exit
+from .setup import config, ssp_exit
 from .ssp_event import SSPEvent
 from .ssp_pick import SSPPick
 logger = logging.getLogger(__name__.rsplit('.', maxsplit=1)[-1])
