+++ sourcespec2/ssp_wave_arrival.py
@@ -16,7 +16,7 @@ import logging
 import warnings
 from math import asin, degrees
 from obspy.taup import TauPyModel
-from .config import config
+from .setup import config
 model = TauPyModel(model='iasp91')
 logger = logging.getLogger(__name__.rsplit('.', maxsplit=1)[-1])
